Alert to rate 5 movies

+++ app/src/main/java/com/chi_square/movierating/data/MovieList.java
@@ -36,7 +36,7 @@ public class MovieList {
         return rating;
     }
 
-    public void setRating(int rating) {
+    public void setRating(float rating) {
         this.rating = rating;
     }
